feat(about): anchor 스크롤 기능 공통화
