+++ src/layout/about/ContentTitle.tsx
@@ -1,11 +1,12 @@
 import styled from '@emotion/styled';
-import React, { useRef } from 'react';
+import React, { useEffect, useRef } from 'react';
 import TextDefault from '@/components/ui/TextDefault';
 import { fontWeight } from '@/utils/StyleTheme';
+import { useAnchor } from '@/hooks/useAnchor';
 
 interface ContentTitleProps {
+  id: string;
   title: string;
-  link: string;
 }
 
 const TitleWrapper = styled.div`
@@ -14,28 +15,15 @@ const TitleWrapper = styled.div`
   border-bottom: 1px solid #9c9c9c;
 `;
 
-export const ContentTitle = ({ title, link }: ContentTitleProps) => {
+export const ContentTitle = ({ title, id }: ContentTitleProps) => {
   const ref = useRef<HTMLAnchorElement | null>(null);
-
-  const handleClickAnchor = () => {
-    if (ref.current) {
-      const refRect = ref.current.getBoundingClientRect().top;
-      const bodyRect = document.body.getBoundingClientRect().top;
-      const navRect = 66 + 16;
-      const top = refRect - bodyRect - navRect;
-
-      window.scrollTo({
-        top,
-        behavior: 'smooth',
-      });
-    }
-  };
+  const { handleClickAnchor } = useAnchor(id);
 
   return (
     <>
       <TitleWrapper>
         <TextDefault size="xxg" weight={fontWeight.medium} color="orange">
-          <a ref={ref} href={'#' + link} onClick={handleClickAnchor}>
+          <a ref={ref} className={id} href={'#' + id} onClick={handleClickAnchor}>
             {title}
           </a>
         </TextDefault>
